:recycle:refactor: main.basic 파일 handleCalculateCartStuff 함수 비즈니스 로직과 UI 로직 분리 및 components, hooks로 구조 변경
